refactor: 불 필요한 모델 타입 제거 및 중첩 타입으로 변경
